+++ apps/3002-checkout/next.config.js
@@ -30,8 +30,8 @@ const nextConfig = {
           './menu': './components/menu',
         },
         shared: {
-          lodash: {},
-          antd: {},
+          'lodash/': {},
+          'antd/': {},
         },
         extraOptions: {
           exposePages: true,